% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continentAsia.R
\name{continentAsia}
\alias{continentAsia}
\title{This function takes slice of data from gp2007}
\usage{
continentAsia()
}
\description{
This function takes slice of data from gp2007
}
\examples{
continentAsia()
}
