% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d}
\alias{d}
\title{Example data from HW 1}
\format{An object of class \code{data.frame} with 30 rows and 2 columns.}
\usage{
d
}
\description{
Example data from HW 1
}
\references{
\url{http://www.stat.umn.edu/geyer/3701/data/q1p4.txt}

The following data are from Quiz 1 / HW 1 for STAT 3701
The first link is used for gamma distributed data (denoted 'ga_data')
The second link is used forr cauchy distributed data (denoted 'cau_data')
The third link is used for binomial distributed data (denoted 'bin_data')

\url{http://www.stat.umn.edu/geyer/3701/data/q1p3.txt}

\url{http://www.stat.umn.edu/geyer/3701/data/q1p7c.txt}

\url{http://users.stat.umn.edu/~almquist/3811_examples/all_alaska_flights.csv}

Weather Data from Lecture 6 from STAT 3701; a tutorial on ggplot2/dplyr

\url{http://users.stat.umn.edu/~almquist/3811_examples/weather.Rda}

Early January Weather data from Lecture 6 from STAT 3701; a tutorial on ggplot2/dplyr

\url{http://users.stat.umn.edu/~almquist/3811_examples/early_january_weather.Rda}
}
\author{
Charlie Geyer \email{geyer@umn.edu}

Charlie Geyer \email{geyer@umn.edu}

Zack W. Almquist \email{almquist@umn.edu}

Zack W. Almquist \email{almquist@umn.edu}

Zack W. Almquist \email{almquist@umn.edu}
}
\keyword{datasets}
