% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignetteplotgp2007.R
\docType{data}
\name{gp2007}
\alias{gp2007}
\title{This vignette shows 3 ways to plot the data from gp2007}
\format{An object of class \code{data.frame} with 142 rows and 6 columns.}
\usage{
gp2007
}
\description{
This vignette shows 3 ways to plot the data from gp2007
}
\keyword{datasets}
