% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NguyenToolsnguyenPkgRfunctionsHW2.R
\name{\%xAx\%}
\alias{\%xAx\%}
\title{Given a numeric vector x and a numeric matrix A, calculates x'inv(A)x, where ' is transposition
This is computed using the function as a binary operator}
\usage{
a \%xAx\% x
}
\arguments{
\item{x}{numeric vector}

\item{A}{numeric matrix}
}
\value{
scalar number computed by x'inv(A)x
}
\description{
Computes x'inv(A)x of a numeric matrix A and numeric vector x.
}
\examples{
load(url("http://www.stat.umn.edu/geyer/3701/data/q2p1.rda"))
ls()
a <- as.matrix(a)
x <- as.matrix(x)
a \%xAx\% x
}
