% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab5.R
\name{lab5}
\alias{lab5}
\title{This function returns mle}
\usage{
lab5(X, density, interval)
}
\description{
This function returns mle
}
\examples{
ga_data <- scan(url('http://www.stat.umn.edu/geyer/3701/data/q1p3.txt'))
ga <- function(theta, x) dgamma(x, shape = theta, log = TRUE)
lab5(ga_data,ga,c(0,3))
}
