% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NguyenToolsnguyenPkgRfunctionsHW2.R
\name{myapply}
\alias{myapply}
\title{Given a matrix X, an integer MARGIN in c(1,2) indicating 1 for rows, and 2 for columns, and a function FUN
Returns the output of what the base R function apply() would do to the matrix X}
\usage{
myapply(X, MARGIN, FUN, ...)
}
\arguments{
\item{X}{a matrix of any type}

\item{MARGIN}{a binary value taking on '1' for rows or '2' for columns of a matrix}

\item{FUN}{a function that operates along the MARGIN of the matrix X}
}
\value{
The output that is identical to what apply() function family would return
}
\description{
Runs a function FUN along the MARGIN of a matrix X
}
\examples{
fred <- matrix(1:6, ncol = 2)
myapply(fred, 1, mean)
myapply(fred, 2, mean)
myapply(fred, 1, max)
myapply(fred, 2, max)
myapply(fred, 1, function(x) quantile(x, 0.75))
myapply(fred, 2, function(x) quantile(x, 0.75))
}
