% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NguyenToolsnguyenPkgRfunctionsHW2.R
\name{p6wrapper}
\alias{p6wrapper}
\title{A wrapper that returns solutions to problem 6 of HW2 from STAT 3701}
\usage{
p6wrapper()
}
\description{
Wrapper function that returns output of the median function along margins of a 3D dataset
}
\examples{
p6wrapper()
}
