% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NguyenToolsnguyenPkgRfunctionsHW2.R
\name{standardize}
\alias{standardize}
\title{Given a numeric matrix, the following 'standardize' function standardizes the columns of the matrix}
\usage{
standardize(a)
}
\arguments{
\item{a}{numeric matrix}
}
\value{
Standardized matrix
}
\description{
Standardizes the columns of a numeric matrix. Standardization is computed as (value - mean(col))/(sd(col))
}
\examples{
load(url("http://www.stat.umn.edu/geyer/3701/data/q2p3.rda"))
ls()
standardize(a)
}
